refactor: modularize voting session page for improved maintainability
