+++ app/users/page.tsx
@@ -1,7 +1,7 @@
 "use client"
 
 import { useEffect, useState } from "react"
-import { getUsers } from "@/app/actions/users"
+import { users as usersActions } from "@/app/actions/index"
 import type { User } from "@/types"
 import { Card, CardContent } from "@/components/ui/card"
 import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert"
@@ -17,7 +17,7 @@ export default function UsersPage() {
   useEffect(() => {
     async function loadUsers() {
       try {
-        const result = await getUsers()
+        const result = await usersActions.getUsers()
         if (result.error) {
           setError(result.error.message)
         } else if (result.data) {
